+++ kedro/validation/__init__.py
@@ -2,15 +2,18 @@
 
 from .core import (
     CallableValidator,
-    CheckFailure,
-    DataValidationError,
-    ValidationConfigurationError,
     Validator,
     ValidatorSpec,
     preflight_check,
     resolve_validator,
 )
-from .exceptions import ModelInstantiationError, ParameterValidationError
+from .exceptions import (
+    CheckFailure,
+    DataValidationError,
+    ModelInstantiationError,
+    ParameterValidationError,
+    ValidationConfigurationError,
+)
 from .model_factory import instantiate_model
 from .pandera_validator import PanderaValidator
 from .parameter_validator import ParameterValidator
